+++ MatrixSDK/MXRestClient.m
@@ -50,6 +50,11 @@ NSString *const kMXRoomVisibilityPrivate = @"private";
 NSString *const kMX3PIDMediumEmail  = @"email";
 NSString *const kMX3PIDMediumMSISDN = @"msisdn";
 
+/**
+ MXRestClient error domain
+ */
+NSString *const kMXRestClientErrorDomain = @"kMXRestClientErrorDomain";
+
 
 /**
  Authentication flow: register or login
@@ -294,6 +299,8 @@ failure:(void (^)(NSError *error))failure
                                   append:(BOOL)append
                                  success:(void (^)())success
                                  failure:(void (^)(NSError *))failure {
+    // Fill the request parameters on demand
+    // Caution: parameters are JSON serialized in http body, we must use a NSNumber created with a boolean for append value.
     NSDictionary *parameters = @{
                                  @"pushkey": pushkey,
                                  @"kind": kind,
@@ -331,6 +338,167 @@ failure:(void (^)(NSError *error))failure
                                  }];
 }
 
+- (MXHTTPOperation *)enablePushRule:(NSString*)ruleId
+                              scope:(NSString*)scope
+                               kind:(MXPushRuleKind)kind
+                             enable:(BOOL)enable
+                            success:(void (^)())success
+                            failure:(void (^)(NSError *error))failure
+{
+    NSString *kindString;
+    switch (kind)
+    {
+        case MXPushRuleKindOverride:
+            kindString = @"override";
+            break;
+        case MXPushRuleKindContent:
+            kindString = @"content";
+            break;
+        case MXPushRuleKindRoom:
+            kindString = @"room";
+            break;
+        case MXPushRuleKindSender:
+            kindString = @"sender";
+            break;
+        case MXPushRuleKindUnderride:
+            kindString = @"underride";
+            break;
+    }
+    
+    NSDictionary *headers = @{@"Content-Type": @"application/json"};
+    
+    NSString *enabled = enable ? @"true": @"false";
+    
+    return [httpClient requestWithMethod:@"PUT"
+                                    path:[NSString stringWithFormat:@"v1/pushrules/%@/%@/%@/enabled", scope, kindString, ruleId]
+                              parameters:nil
+                                    data:[enabled dataUsingEncoding:NSUTF8StringEncoding]
+                                 headers:headers
+                                 timeout:-1
+                          uploadProgress:nil
+                                 success:^(NSDictionary *JSONResponse) {
+                                     if (success)
+                                     {
+                                         success();
+                                     }
+                                 }
+                                 failure:^(NSError *error) {
+                                     if (failure)
+                                     {
+                                         failure(error);
+                                     }
+                                 }];
+}
+
+- (MXHTTPOperation *)removePushRule:(NSString*)ruleId
+                              scope:(NSString*)scope
+                               kind:(MXPushRuleKind)kind
+                            success:(void (^)())success
+                            failure:(void (^)(NSError *error))failure
+{
+    NSString *kindString;
+    switch (kind)
+    {
+        case MXPushRuleKindOverride:
+            kindString = @"override";
+            break;
+        case MXPushRuleKindContent:
+            kindString = @"content";
+            break;
+        case MXPushRuleKindRoom:
+            kindString = @"room";
+            break;
+        case MXPushRuleKindSender:
+            kindString = @"sender";
+            break;
+        case MXPushRuleKindUnderride:
+            kindString = @"underride";
+            break;
+    }
+    
+    return [httpClient requestWithMethod:@"DELETE"
+                                    path:[NSString stringWithFormat:@"v1/pushrules/%@/%@/%@", scope, kindString, ruleId]
+                              parameters:nil
+                                 success:^(NSDictionary *JSONResponse) {
+                                     if (success)
+                                     {
+                                         success();
+                                     }
+                                 }
+                                 failure:^(NSError *error) {
+                                     if (failure)
+                                     {
+                                         failure(error);
+                                     }
+                                 }];
+}
+
+- (MXHTTPOperation *)addPushRule:(NSString*)ruleId
+                           scope:(NSString*)scope
+                            kind:(MXPushRuleKind)kind
+                         actions:(NSArray*)actions
+                         pattern:(NSString*)pattern
+                         success:(void (^)())success
+                         failure:(void (^)(NSError *error))failure
+{
+    NSString *kindString;
+    NSDictionary *content = nil;
+    
+    switch (kind)
+    {
+        case MXPushRuleKindContent:
+            kindString = @"content";
+            if (pattern.length && actions.count)
+            {
+                content = @{@"pattern": pattern, @"actions": actions};
+            }
+            break;
+        case MXPushRuleKindRoom:
+            kindString = @"room";
+            if (actions.count)
+            {
+                content = @{@"actions": actions};
+            }
+            break;
+        case MXPushRuleKindSender:
+            kindString = @"sender";
+            if (actions.count)
+            {
+                content = @{@"actions": actions};
+            }
+            break;
+        default:
+            break;
+    }
+
+    // Sanity check
+    if (content)
+    {
+        return [httpClient requestWithMethod:@"PUT"
+                                        path:[NSString stringWithFormat:@"v1/pushrules/%@/%@/%@", scope, kindString, ruleId]
+                                  parameters:content
+                                     success:^(NSDictionary *JSONResponse) {
+                                         if (success)
+                                         {
+                                             success();
+                                         }
+                                     }
+                                     failure:^(NSError *error) {
+                                         if (failure)
+                                         {
+                                             failure(error);
+                                         }
+                                     }];
+    }
+    else
+    {
+        if (failure)
+        {
+            failure([NSError errorWithDomain:kMXRestClientErrorDomain code:0 userInfo:@{@"error": @"Invalid argument"}]);
+        }
+        return nil;
+    }
+}
 
 #pragma mark - Room operations
 - (MXHTTPOperation*)sendEventToRoom:(NSString*)roomId
@@ -799,9 +967,9 @@ failure:(void (^)(NSError *error))failure
 {
     NSString *path = [NSString stringWithFormat:@"v1/rooms/%@/typing/%@", roomId, self.credentials.userId];
     
-    // All query parameters are optional. Fill the request parameters on demand
+    // Fill the request parameters on demand
     NSMutableDictionary *parameters = [NSMutableDictionary dictionary];
-    
+    // Caution: parameters are JSON serialized in http body, we must use a NSNumber created with a boolean for typing value.
     parameters[@"typing"] = [NSNumber numberWithBool:typing];
     
     if (-1 != timeout)